src/species-classification/requirements.txt
torch==1.9.0
torchvision==0.10.0
pillow==9.5.0
pandas
numpy
scikit-learn
google-cloud-storage
google-auth
tensorflow

